fix: rename orpc util
