+++ apps/web/src/components/query-client.provider.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { QueryClientProvider as TanStackQueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "@/utils/orcp";
+import { queryClient } from "@/utils/orpc";
 import type { ReactNode } from "react";
 
 export default function QueryClientProvider({
